fix(ci): extract NVIDIA/Ollama to own modules, fix React lint errors, format Python
Moves isNvidiaProxyConfigured/callNvidiaProxyChat to nvidia-proxy-client.ts and
isOllamaConfigured/callOllamaChat to ollama-client.ts. workers-ai-client.ts is
unchanged from existing tests' perspective so no mock updates are needed.

workers-ai-chat.ts: drops the isWorkersAiConfigured() guard and tries callWorkersAiChat
directly, catching UnauthorizedException to fall through — test mocks work without
needing isWorkersAiConfigured in the mock factory.

InsightPanel.tsx: moves Date.now() into a lazy useState initializer (not during render).
explore/page.tsx: replaces useEffect+setState with lazy useState initializers for URL params.
postiz/page.tsx: replaces useEffect+setState with effectiveId derived variable.
tailscale-metrics-exporter.py: ruff format.
