+++ src/lib/workers-ai-client.ts
@@ -124,166 +124,6 @@ export async function callWorkersAiChat(
   }
 }
 
-// ---------------------------------------------------------------------------
-// NVIDIA proxy chat (nemotron-3.5-lightning-30b-a3b with thinking)
-// ---------------------------------------------------------------------------
-
-export function isNvidiaProxyConfigured(): boolean {
-  return Boolean(process.env.NVIDIA_PROXY_URL && process.env.NVIDIA_PROXY_TOKEN);
-}
-
-/**
- * Call the postiz-ai-proxy Worker with ?thinking=1 to invoke nemotron-3.5-lightning
- * with extended thinking. reasoning_content is stripped server-side in the Worker;
- * this function receives only the final assistant content.
- */
-export async function callNvidiaProxyChat(
-  messages: { role: string; content: string }[],
-  opts?: { maxTokens?: number }
-): Promise<string> {
-  const proxyUrl = process.env.NVIDIA_PROXY_URL;
-  const proxyToken = process.env.NVIDIA_PROXY_TOKEN;
-  if (!proxyUrl || !proxyToken) {
-    const err = new Error("NVIDIA proxy not configured");
-    err.name = "UnauthorizedException";
-    throw err;
-  }
-
-  const started = Date.now();
-  try {
-    const response = await fetch(`${proxyUrl}/chat/completions?thinking=1`, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${proxyToken}`,
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        messages,
-        max_tokens: opts?.maxTokens ?? 600,
-        temperature: 0.6,
-        top_p: 0.95,
-      }),
-    });
-
-    if (!response.ok) {
-      const text = await response.text();
-      const err = new Error(`NVIDIA proxy HTTP ${response.status}: ${text.slice(0, 200)}`);
-      err.name =
-        response.status === 401 || response.status === 403 ? "UnauthorizedException" : "AiError";
-      recordAdminAiCall({
-        ok: false,
-        model: "nvidia/nemotron-3.5-lightning-30b-a3b",
-        viaGateway: false,
-        latencyMs: Date.now() - started,
-        error: err.message,
-      });
-      throw err;
-    }
-
-    const data = (await response.json()) as {
-      choices?: Array<{ message?: { content?: string } }>;
-    };
-    const text = data.choices?.[0]?.message?.content?.trim() ?? "";
-    recordAdminAiCall({
-      ok: true,
-      model: "nvidia/nemotron-3.5-lightning-30b-a3b",
-      viaGateway: false,
-      latencyMs: Date.now() - started,
-    });
-    return text || "Sorry — I could not generate a reply.";
-  } catch (err) {
-    if (err instanceof Error && (err.name === "UnauthorizedException" || err.name === "AiError")) {
-      throw err;
-    }
-    recordAdminAiCall({
-      ok: false,
-      model: "nvidia/nemotron-3.5-lightning-30b-a3b",
-      viaGateway: false,
-      latencyMs: Date.now() - started,
-      error: err instanceof Error ? err.message : "NVIDIA proxy request failed",
-    });
-    throw err;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Ollama local inference (OpenAI-compatible, self-hosted)
-// ---------------------------------------------------------------------------
-
-export function isOllamaConfigured(): boolean {
-  return Boolean(process.env.OLLAMA_BASE_URL);
-}
-
-/**
- * Call a local Ollama instance via its OpenAI-compatible endpoint.
- * OLLAMA_BASE_URL   — e.g. http://100.121.191.7:11434 (Tailscale IP of WSL host)
- * OLLAMA_MODEL      — optional override; defaults to qwen2.5-coder:latest
- */
-export async function callOllamaChat(
-  messages: { role: string; content: string }[],
-  opts?: { maxTokens?: number }
-): Promise<string> {
-  const baseUrl = process.env.OLLAMA_BASE_URL;
-  if (!baseUrl) {
-    const err = new Error("Ollama not configured");
-    err.name = "UnauthorizedException";
-    throw err;
-  }
-  const model = process.env.OLLAMA_MODEL || "qwen2.5-coder:latest";
-  const started = Date.now();
-
-  try {
-    const response = await fetch(`${baseUrl}/v1/chat/completions`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        model,
-        messages,
-        max_tokens: opts?.maxTokens ?? 600,
-        stream: false,
-      }),
-    });
-
-    if (!response.ok) {
-      const text = await response.text();
-      const err = new Error(`Ollama HTTP ${response.status}: ${text.slice(0, 200)}`);
-      err.name = "AiError";
-      recordAdminAiCall({
-        ok: false,
-        model,
-        viaGateway: false,
-        latencyMs: Date.now() - started,
-        error: err.message,
-      });
-      throw err;
-    }
-
-    const data = (await response.json()) as {
-      choices?: Array<{ message?: { content?: string } }>;
-    };
-    const text = data.choices?.[0]?.message?.content?.trim() ?? "";
-    recordAdminAiCall({
-      ok: true,
-      model,
-      viaGateway: false,
-      latencyMs: Date.now() - started,
-    });
-    return text || "Sorry — I could not generate a reply.";
-  } catch (err) {
-    if (err instanceof Error && (err.name === "UnauthorizedException" || err.name === "AiError")) {
-      throw err;
-    }
-    recordAdminAiCall({
-      ok: false,
-      model,
-      viaGateway: false,
-      latencyMs: Date.now() - started,
-      error: err instanceof Error ? err.message : "Ollama request failed",
-    });
-    throw err;
-  }
-}
-
 export type WorkersAiToolCall = { name: string; args: Record<string, unknown> };
 
 /**
